separated edit vaccination dialog to separate file
